django/requirements/base.txt
Django==4.2.*
django-cleanup==8.0.*
django-cors-headers==4.2.*
django-environ==0.11.*
django-security==0.12.*
djangorestframework==3.14.*
python-dotenv==1.0.*
